requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
typing-extensions>=4.8.0
psycopg2-binary==2.9.9
